refactor(P4-2): .repository サフィックス統一・Prisma接頭辞除去・IF接頭辞（Batch5b）
- ドット区切り *.repository.ts → *-repository.ts（idashboard-query / igeppo /
  iimport-job / iwork-record / dashboard-query / milestone）
- Prisma 接頭辞の除去（ファイル名＋クラス名）:
  - geppo-prisma.repository.ts → geppo-repository.ts / GeppoPrismaRepository → GeppoRepository
  - import-job-prisma.repository.ts → import-job-repository.ts / ImportJobPrismaRepository → ImportJobRepository
  - work-record-prisma.repository.ts → work-record-repository.ts / WorkRecordPrismaRepository → WorkRecordRepository
- リポジトリ IF の接頭辞統一: wbs-query-repository.ts → iwbs-query-repository.ts

本番型チェッククリーン。
