+++ src/applications/dashboard/repositories/idashboard-query-repository.ts
