+++ src/applications/wbs/query/get-wbs-summary-handler.ts
@@ -4,8 +4,8 @@ import { GetWbsSummaryQuery } from "./get-wbs-summary-query";
 import { WbsSummaryResult, PhaseSummary, AssigneeSummary, TaskAllocationDetail, MonthlyAssigneeSummary } from "./wbs-summary-result";
 import { AllocationCalculationMode } from "./allocation-calculation-mode";
 import { SYMBOL } from "@/types/symbol";
-import { WbsTaskData, PhaseData, TaskActualMonthly } from "@/applications/wbs/query/wbs-query-repository";
-import type { IWbsQueryRepository } from "@/applications/wbs/query/wbs-query-repository";
+import { WbsTaskData, PhaseData, TaskActualMonthly } from "@/applications/wbs/query/iwbs-query-repository";
+import type { IWbsQueryRepository } from "@/applications/wbs/query/iwbs-query-repository";
 import { WorkingHoursAllocationService } from "@/domains/calendar/working-hours-allocation-service";
 import { CompanyCalendar } from "@/domains/calendar/company-calendar";
 import type { ICompanyHolidayRepository } from "@/applications/calendar/icompany-holiday-repository";
